requirements.txt
python-telegram-bot>=21,<22
fastapi==0.111.0
uvicorn[standard]==0.30.0
